fix(locale): rename city_prefix to city_name for en_NG, en_ZA, en_AU_ocker
similar to #2109 for en_* locales
