+++ src/locales/en_AU_ocker/location/city_pattern.ts
@@ -1,1 +1,1 @@
-export default ['{{location.city_prefix}}'];
+export default ['{{location.city_name}}'];
