+++ src/locales/en_AU_ocker/location/index.ts
@@ -4,8 +4,8 @@
  */
 import type { LocationDefinitions } from '../../..';
 import building_number from './building_number';
+import city_name from './city_name';
 import city_pattern from './city_pattern';
-import city_prefix from './city_prefix';
 import default_country from './default_country';
 import postcode from './postcode';
 import region from './region';
@@ -17,8 +17,8 @@ import street_suffix from './street_suffix';
 
 const location: LocationDefinitions = {
   building_number,
+  city_name,
   city_pattern,
-  city_prefix,
   default_country,
   postcode,
   region,
