+++ src/locales/en_ZA/location/city_pattern.ts
@@ -1,1 +1,1 @@
-export default ['{{location.city_prefix}}'];
+export default ['{{location.city_name}}'];
