+++ src/locales/en_ZA/location/index.ts
@@ -3,16 +3,16 @@
  * Run 'pnpm run generate:locales' to update.
  */
 import type { LocationDefinitions } from '../../..';
+import city_name from './city_name';
 import city_pattern from './city_pattern';
-import city_prefix from './city_prefix';
 import default_country from './default_country';
 import postcode from './postcode';
 import state from './state';
 import street_pattern from './street_pattern';
 
 const location: LocationDefinitions = {
+  city_name,
   city_pattern,
-  city_prefix,
   default_country,
   postcode,
   state,
